Restructuring code for the IDS crawler to remove dependencies on turnkey linux and host configuration.
The code is currently structured into two distinct directories:

- bootstrap/
- crawler/

bootstrap is concerned with getting the system into a state where it can
run, whilst crawler is concerned with running eldis & r4d crawls and
ultimately importing the crawled data into a triple store.

The code is currently broken and needs fixed, but it was copied here
from the following github repo & commit:

    repo: https://github.com/neontribe/Linked_Development
  commit: 5d16ef2148d1796c229049372e89a95509e4e07c

This should all be cleaned up further and made to work in subsequent commits.

At some point we may add a Dockerfile to formalise this further into a
repeatable build and deployment process.